// src/modules/auth/domain/auth.types.ts
// auth.types.ts
import { UserRole } from '@/modules/auth/domain/user.entity';

export interface AuthUser {
  id: string;
  tenantId?: string | null;
  email: string;
  roles: UserRole[];
  iat: number;
  exp: number;
}

export interface JwtAuthPayload {
  sub: string;
  email: string;
  roles: UserRole[];
  iat: number;
  exp: number;
}

export interface CookieType {
  auth_token: {
    set: (options: {
      value: string;
      httpOnly: boolean;
      secure: boolean;
      sameSite: 'strict' | 'lax' | 'none';
      maxAge: number;
      path: string;
    }) => void;
  };
}